Added resizing via target render texture
fps is dropping a bit when using render texture but it's tolerable

+++ src/declaration.cpp
@@ -6,8 +6,8 @@ int window_width = 1280; // maybe '_' (internal) too
 int window_height = 720;
 
 //initial "source" screen resolution - maybe move to global
-const float initial_window_width = window_width;
-const float initial_window_height = window_height;
+const int initial_window_width = window_width;
+const int initial_window_height = window_height;
 
 float scale_x = 1.0f; // scales should be global i guess
 float scale_y = 1.0f;
